new updates

+++ lib/calculas.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
 
-double df(double (*func)(int),double x){
-    // double eps = 1.61011979e-4;
-    double eps = 1.61012e-4;
-    double df = (func(x+eps) -func(x-eps))/(2*eps);
+float df(float (*func)(float),float x){
+    // float eps = 1.61011979e-4;
+    float eps = 1.61012e-4;
+    // float df = (func(x+eps) -func(x-eps))/(2*eps);
+    float df = (func(x+eps) -func(x-eps))/(2*eps);
     return df;
 }
 
@@ -11,13 +12,13 @@ double df(double (*func)(int),double x){
 
 /*
 
-double func(double x){
+float func(float x){
     return x*x;
 }
 
 int main(){
 
-    double dff = df(func,3);
+    float dff = df(func,3);
 
     printf("%f\n",dff);
     return 0;
